Replace RegexTools with one file per function
