+++ lib/getUniqueMatch.js
@@ -2,20 +2,6 @@
 var Cesium = require('cesium');
 var defined = Cesium.defined;
 
-var RegexTools = {};
-module.exports = RegexTools;
-
-var googleLinkRegex = /https?:\/\/groups\.google\.com[^\s.,:]*/ig;
-
-/** Find unique GoogleGroup links in array of strings
- *
- * @param {String[]} textArray Array of strings to regex search
- * @return {String[]} Unique links
- */
-RegexTools.getGoogleGroupLinks = function(textArray) {
-    return RegexTools._getUniqueMatch(textArray, googleLinkRegex);
-};
-
 /** Generic regex search over array of String. (Requires global regex)
  *
  * @private
@@ -23,7 +9,7 @@ RegexTools.getGoogleGroupLinks = function(textArray) {
  * @param {Object} regex Global Regex
  * @returns {String[]} Unique matches of the regex
  */
-RegexTools._getUniqueMatch = function(textArray, regex) {
+module.exports = function getUniqueMatch(textArray, regex) {
     var matches = [];
     var matchResult;
     if (!defined(textArray) || !defined(regex)) {
